abc389

+++ atcoder/contest/abc389/c/main.cpp
@@ -111,14 +111,14 @@ using namespace std;
 int main() {
     int q;
     cin >> q;
-    long long now = 0;
-    vector<long long> x;
+    ll now = 0;
+    vector<ll> x;
     int id = 0;
     for (int i = 0; i < q; ++i) {
         int t;
         cin >> t;
         if (t == 1) {
-            long long l;
+            ll l;
             cin >> l;
             x.push_back(now);
             now += l;
